updated external comfort methods
