+++ LadybugTools_Engine/Python/src/ladybugtools_toolkit/external_comfort/external_comfort.py
@@ -23,7 +23,6 @@ from ..bhomutil.encoder import (
     inf_str_to_inf_dtype,
     pascalcase,
 )
-from ..external_comfort.utci import categorise, utci_comfort_categories
 from ..helpers import evaporative_cooling_effect
 from ..ladybug_extension.analysis_period import (
     AnalysisPeriod,
@@ -49,6 +48,7 @@ from ..plot.utci_heatmap_histogram import utci_heatmap_histogram
 from .shelter import Shelter
 from .simulate import SimulationResult
 from .typology import Typology
+from .utci import categorise, utci_comfort_categories
 from .utci.calculate import utci
